cert: Switch the certs to be symlink based
Actual certs are stored in certs/YEAR/name-DATE.*

Canonical symlinks are certs/name.* and point to the actual cert.

Symlinks are only updated if cert process completes with no
erros. Otherwise they remain pointed to the old certs. Restarting on
error still is Make-like in that it won't recreate things (as long as
it's the same date).